test(guardrails): fix OpenTelemetry operator cleanup logic and consolidate HuggingFace detector tests
  - Fix installed_opentelemetry_operator fixture to properly handle post-upgrade cleanup
  - Add post_upgrade conditional logic to vllm_cpu_runtime fixture for resource reuse
  - Consolidate guardrails HuggingFace detector upgrade tests into test_guardrails_upgrade.py
  - Remove duplicate test_guardrails_upgrades.py file
  - Remove unnecessary comment from prompt_injection_detector_isvc fixture
  - Remove conditional teardown guard from installed_tempo_operator fixture
