+++ tests/fixtures/inference.py
@@ -36,25 +36,38 @@ def vllm_cpu_runtime(
     minio_pod: Pod,
     minio_service: Service,
     minio_data_connection: Secret,
+    pytestconfig: pytest.Config,
+    teardown_resources: bool,
 ) -> Generator[ServingRuntime, Any, Any]:
-    with ServingRuntimeFromTemplate(
-        client=admin_client,
-        name="vllm-runtime-cpu-fp16",
-        namespace=model_namespace.name,
-        template_name=RuntimeTemplates.VLLM_CUDA,
-        deployment_type=KServeDeploymentType.RAW_DEPLOYMENT,
-        runtime_image="quay.io/rh-aiservices-bu/vllm-cpu-openai-ubi9"
-        "@sha256:ada6b3ba98829eb81ae4f89364d9b431c0222671eafb9a04aa16f31628536af2",
-        containers={
-            "kserve-container": {
-                "args": ["--port=8032", "--model=/mnt/models", "--served-model-name={{.Name}}"],
-                "ports": [{"containerPort": 8032, "protocol": "TCP"}],
-                "volumeMounts": [{"mountPath": "/dev/shm", "name": "shm"}],
-            }
-        },
-        volumes=[{"emptyDir": {"medium": "Memory", "sizeLimit": "2Gi"}, "name": "shm"}],
-    ) as serving_runtime:
+    if pytestconfig.option.post_upgrade:
+        # During post-upgrade, reuse existing ServingRuntime
+        serving_runtime = ServingRuntime(
+            client=admin_client,
+            name="vllm-runtime-cpu-fp16",
+            namespace=model_namespace.name,
+        )
         yield serving_runtime
+        if teardown_resources:
+            serving_runtime.clean_up()
+    else:
+        with ServingRuntimeFromTemplate(
+            client=admin_client,
+            name="vllm-runtime-cpu-fp16",
+            namespace=model_namespace.name,
+            template_name=RuntimeTemplates.VLLM_CUDA,
+            deployment_type=KServeDeploymentType.RAW_DEPLOYMENT,
+            runtime_image="quay.io/rh-aiservices-bu/vllm-cpu-openai-ubi9"
+            "@sha256:ada6b3ba98829eb81ae4f89364d9b431c0222671eafb9a04aa16f31628536af2",
+            containers={
+                "kserve-container": {
+                    "args": ["--port=8032", "--model=/mnt/models", "--served-model-name={{.Name}}"],
+                    "ports": [{"containerPort": 8032, "protocol": "TCP"}],
+                    "volumeMounts": [{"mountPath": "/dev/shm", "name": "shm"}],
+                }
+            },
+            volumes=[{"emptyDir": {"medium": "Memory", "sizeLimit": "2Gi"}, "name": "shm"}],
+        ) as serving_runtime:
+            yield serving_runtime
 
 
 @pytest.fixture(scope="class")
